Controller/pyproject.toml
[project]
name = "controller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "debugpy>=1.8.14",
    "matplotlib>=3.10.1",
    "numpy>=2.2.3",
    "opencv-python>=4.11.0.86",
    "paho-mqtt>=2.1.0",
    "pathfinding>=1.0.16",
    "ruff>=0.9.9",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
]
